+++ CHANGELOG.md
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.6.0] - 2023-10-29
+### Added
+- TOC: New option `slugify` to configure how the ids are generated
+  - It can be a function `(value: string) => string`
+  - Or an object with the options for the `createSlugifier` function.
+  If it's not set, a new slugifier is created with the default options.
+
+### Changed
+- TOC plugin requires Lume v1.19.3
+
 ## [0.5.1] - 2023-08-01
 ### Fixed
 - `footnotes` plugin attributes.
@@ -37,6 +47,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - `toc` plugin.
 - `title` plugin.
 
+[0.6.0]: https://github.com/lumeland/markdown-plugins/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/lumeland/markdown-plugins/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/lumeland/markdown-plugins/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/lumeland/markdown-plugins/compare/v0.3.0...v0.4.0